Replace gunicorn with simple uvicorn without multiprocessing
The way the app is written does not really work with multiple processes.
Normally an ASGI app should be scalable by just creating multiple
server processes, but here we run an expensive task in the background,
which I believe then runs in *every* process - that's not what we want.
Let's stick with a very simple, single process uvicorn (and derive from
the normal Python Docker image).